fix: extract types from popup
