+++ src/packages/cascader/cascader.taro.tsx
@@ -10,7 +10,7 @@ import React, {
 import classNames from 'classnames'
 import { Loading, Check } from '@nutui/icons-react-taro'
 import { ScrollView, View } from '@tarojs/components'
-import { Popup, PopupProps } from '@/packages/popup/popup.taro'
+import Popup, { PopupProps } from '@/packages/popup/index.taro'
 import { Tabs } from '@/packages/tabs/tabs.taro'
 import Tree, { convertListToOptions } from './utils'
 import {
