+++ src/packages/numberkeyboard/numberkeyboard.taro.tsx
@@ -2,8 +2,7 @@ import React, { FunctionComponent, ReactNode, useMemo, useState } from 'react'
 import classNames from 'classnames'
 import { ArrowDown } from '@nutui/icons-react-taro'
 import { View } from '@tarojs/components'
-import Popup from '@/packages/popup/index.taro'
-import { PopupProps } from '@/packages/popup/popup.taro'
+import Popup, { PopupProps } from '@/packages/popup/index.taro'
 import { useConfig } from '@/packages/configprovider/index.taro'
 import { ComponentDefaults } from '@/utils/typings'
 import { mergeProps } from '@/utils/merge-props'
