+++ src/packages/popup/index.taro.ts
@@ -1,4 +1,4 @@
 import { Popup } from './popup.taro'
 
-export type { PopupProps } from './popup.taro'
+export type { PopupProps } from './types'
 export default Popup
