+++ src/packages/popup/index.ts
@@ -1,4 +1,4 @@
 import { Popup } from './popup'
 
-export type { PopupProps } from './popup'
+export type { PopupProps } from './types'
 export default Popup
