+++ src/packages/popup/types.ts
@@ -1,0 +1,24 @@
+import React, { ReactNode } from 'react'
+import { OverlayProps } from '@/packages/overlay/types'
+
+export type Teleport = HTMLElement | (() => HTMLElement) | null
+export interface PopupProps extends OverlayProps {
+  position: string
+  transition: string
+  overlayStyle: React.CSSProperties
+  overlayClassName: string
+  closeable: boolean
+  closeIconPosition: string
+  closeIcon: ReactNode
+  left?: ReactNode
+  title?: ReactNode
+  description?: ReactNode
+  destroyOnClose: boolean
+  portal: Teleport
+  overlay: boolean
+  round: boolean
+  onOpen: () => void
+  onClose: () => void
+  onOverlayClick: (e: TouchEvent) => boolean | void
+  onCloseIconClick: (e: TouchEvent) => boolean | void
+}
